moar RxJava tweaks
- replaced the map operator that pulls out the video Id’s from the
playlistItemsListResponse with a flattop operator that emits an
Observable which in turn maps each playlistItem to the videoId then
turns the result back into a list
- updated the filter operator so that we return true if the
videoListItems list is not null and the list size is not 0.  returning
true means that we want to keep it in the stream.
- fixed a couple of minor warnings. Made mLayoutManager a local
variable and added javadoc for the youTubeDataApi method parameter